+++ VComponents/Components/Containers/Sheet/VSheet.swift
@@ -24,7 +24,6 @@ import SwiftUI
 ///                 VList(
 ///                     layout: .fixed,
 ///                     data: 1..<10,
-///                     id: \.self,
 ///                     rowContent: { num in
 ///                         Text(String(num))
 ///                             .padding(.vertical, 2)
@@ -100,7 +99,6 @@ struct VSheet_Previews: PreviewProvider {
                 VList(
                     layout: .fixed,
                     data: 1..<10,
-                    id: \.self,
                     rowContent: { num in
                         Text(String(num))
                             .padding(.vertical, 2)